# week04/Demo/api/biohouse.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "biohouse.proto",
        "version": ""
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/Biohouse/ReqAimMove": {
            "post": {
                "summary": "/Biohouse/ReqAimMove",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCAimMove"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSAimMove"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqAttack": {
            "post": {
                "summary": "/Biohouse/ReqAttack",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCAttack"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSAttack"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqAttackNull": {
            "post": {
                "summary": "/Biohouse/ReqAttackNull",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCAttackNull"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSAttackNull"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqChangeEquipWeaponID": {
            "post": {
                "summary": "/Biohouse/ReqChangeEquipWeaponID",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCChangeEquipWeaponID"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSChangeEquipWeaponID"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqChangeHandWeapon": {
            "post": {
                "summary": "/Biohouse/ReqChangeHandWeapon",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCChangeHandWeapon"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSChangeHandWeapon"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqChangeRatio": {
            "post": {
                "summary": "/Biohouse/ReqChangeRatio",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCChangeRatio"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSChangeRatio"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqChangeWeaponPart": {
            "post": {
                "summary": "/Biohouse/ReqChangeWeaponPart",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCChangeWeaponPart"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSChangeWeaponPart"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqEnterTable": {
            "post": {
                "summary": "/Biohouse/ReqEnterTable",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCEnterTable"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSEnterTable"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqHurtPlayer": {
            "post": {
                "summary": "/Biohouse/ReqHurtPlayer",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCHurtPlayer"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSHurtPlayer"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqLeaveTable": {
            "post": {
                "summary": "/Biohouse/ReqLeaveTable",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCLeaveTable"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSLeaveTable"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqLogin": {
            "post": {
                "summary": "/Biohouse/ReqLogin",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCLogin"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSLogin"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqLogout": {
            "post": {
                "summary": "/Biohouse/ReqLogout",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCLogout"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSLogout"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqMaterial": {
            "post": {
                "summary": "/Biohouse/ReqMaterial",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCMaterial"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSMaterial"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
...
        "/Biohouse/ReqPlayerHPChange": {
            "post": {
                "summary": "/Biohouse/ReqPlayerHPChange",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCPlayerHPChange"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSPlayerHPChange"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqUpgradeWeaponPart": {
            "post": {
                "summary": "/Biohouse/ReqUpgradeWeaponPart",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCUpgradeWeaponPart"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSUpgradeWeaponPart"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqUseSkill": {
            "post": {
                "summary": "/Biohouse/ReqUseSkill",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCUseSkill"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSUseSkill"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        },
        "/Biohouse/ReqWeaponReload": {
            "post": {
                "summary": "/Biohouse/ReqWeaponReload",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.biohouse.SCWeaponReload"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.biohouse.CSWeaponReload"
                        }
                    }
                ],
                "tags": [
                    "biohouse.Biohouse"
                ]
            }
        }
    },
    "definitions": {
...
        ".biohouse.CSAimMove": {
            "type": "object",
            "properties": {
                "Point": {
                    "$ref": "#/definitions/.biohouse.PointInfo"
                }
            },
            "description": "准星移动（推送给同桌,自己不需要回包）"
        },
        ".biohouse.CSAttack": {
            "type": "object",
            "properties": {
                "CreateID": {
                    "type": "integer"
                },
                "PartID": {
                    "type": "integer"
                }
            },
            "description": "杀怪"
        },
        ".biohouse.CSAttackNull": {
            "type": "object",
            "properties": {
                "Point": {
                    "$ref": "#/definitions/.biohouse.PointInfo"
                }
            },
            "description": "放空枪，（推送给同桌,自己不需要回包）"
        },
        ".biohouse.CSChangeEquipWeaponID": {
            "type": "object",
            "properties": {
                "EquipID": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            },
            "description": "更换身上的武器,装备栏上的3个位置0、1、2都要发送过来"
        },
        ".biohouse.CSChangeHandWeapon": {
            "type": "object",
            "properties": {
                "HandID": {
                    "type": "integer"
                }
            },
            "description": "更换手上使用的是哪个的武器栏位id"
        },
...
        ".biohouse.CSLeaveTable": {
            "type": "object",
            "properties": {},
            "description": "客户请求离桌请求"
        },
...
        ".biohouse.SCAimMove": {
            "type": "object",
            "properties": {
                "PlayerID": {
                    "type": "integer"
                },
                "Point": {
                    "$ref": "#/definitions/.biohouse.PointInfo"
                }
            },
            "description": "push给桌子中其他玩家 自己不需要"
        },
        ".biohouse.SCAttack": {
            "type": "object",
            "properties": {
                "PlayerID": {
                    "type": "integer"
                },
                "CreateID": {
                    "type": "integer"
                },
                "Money": {
                    "type": "integer"
                },
                "LeftBlood": {
                    "type": "integer"
                },
                "Left": {
                    "$ref": "#/definitions/.biohouse.BloodLeft"
                },
                "Clip": {
                    "type": "integer"
                }
            },
            "description": "push给所有玩家玩家, CSAttack回包为空 后面需要给桌子中所有玩家的都是用push"
        },
        ".biohouse.SCAttackNull": {
            "type": "object",
            "properties": {
                "PlayerID": {
                    "type": "integer"
                },
                "Point": {
                    "$ref": "#/definitions/.biohouse.PointInfo"
                }
            },
            "description": "push给桌子中其他玩家 自己不需要"
        },
        ".biohouse.SCChangeEquipWeaponID": {
            "type": "object",
            "properties": {}
        },
        ".biohouse.SCChangeHandWeapon": {
            "type": "object",
            "properties": {
                "PlayerID": {
                    "type": "integer"
                },
                "HandID": {
                    "type": "integer"
                },
                "Ratio": {
                    "type": "integer"
                }
            },
            "description": "push到桌子中其他玩家 换了武器 则默认的倍率也会使用当前武器装备的品阶倍率"
        },
...
    }
}